write stream data directly into the output packet
This adds support for emitting stream send buffer data directly to the
output buffer.

The idea is that instead of removing data from the send buffer, and
moving it to a new intermediate RangeBuf, which will then be written
to the packet, the data is written directly to the packet buffer.

For retransmission purposes, the data needs to be held in the send
buffer until it is acked, so the first few RangeBuf objects at the start
of the buffer might not actually correspond to the data that needs to be
written next, so a separate index is kept for the whole SendBuf that
indicates which buffer should be written next.

When data is acked, it is marked as such, but only dropped from the fron
of the send buffer in contiguous blocks, to avoid having to move
elements of the VecDeque around in case of gaps.

In addition, due to the fact that we don't know the length of the data
emitted by SendBuf in advance, the CRYPTO and STREAM frames' headers
need to be encoded into the packet after the frames' payload. This is
done by reserving some space before the payload is written.

However since the length is encoded as a variable integer, and we don't
know the value at the time the space is reserved, we always use a 2-byte
varint to encode the length field. The offset is known beforehand so we
can use the most efficient encoding.